Refactor genre states using context
