+++ React/src/user/Navbar.jsx
@@ -13,8 +13,6 @@ import ThumbUp from '@mui/icons-material/ThumbUp';
 
 const Navbar = ({ setSearchQuery, displayMyList, setDisplayMyList, displayWishlist, setDisplayWishlist, userName, setUser, loginModalOpen, setLoginModalOpen, isLoggedIn, setIsLoggedIn }) => {
   const [inputValue, setInputValue] = useState('');
-  const [genres, setGenres] = useState([]);
-  const [likedGenres, setLikedGenres] = useState([]);
 
   //TODO: handleSubmit and handleAutClick is from the odin proj, should be changed.
   const handleSubmit = (e) => {
@@ -41,44 +39,6 @@ const Navbar = ({ setSearchQuery, displayMyList, setDisplayMyList, displayWishli
     }
 
   };
-
-  useEffect(() => {
-    const fetchGenres = async () => {
-      try {
-        //setGenres(getGenres)// TODO: Fix when endpoint implemented
-        getGenres('', 40).then(data => {
-          setGenres(data);
-        })
-      }
-      catch (e) {
-        console.log(e)
-      }
-    };
-
-    fetchGenres();
-  }, []);
-  useEffect(() => {
-    const fetchGenresPrefs = async () => {
-      if (userName != null)
-      {
-        try {
-          getGenrePreferences(userName).then(data => {
-            setLikedGenres(data);
-          })
-        }
-        catch (e) {
-          console.log(e);
-          setLikedGenres([]);
-        }
-      }
-      else
-      {
-        setLikedGenres([]);
-      }
-    };
-
-    fetchGenresPrefs();
-  }, [userName]);
   
   const searchStart = () => {
     alert("Search functinality here (if user presses button instead of pressing Enter")
@@ -108,7 +68,7 @@ const Navbar = ({ setSearchQuery, displayMyList, setDisplayMyList, displayWishli
         >
           {displayWishlist ? <FavoriteIcon/>: <FavoriteBorderIcon/>}
         </button>
-        {userName!=null && <GenreDropdown genres={genres} likedGenres={likedGenres} userName={userName}/>}
+        {userName!=null && <GenreDropdown userName={userName}/>}
         
         <form onSubmit={handleSubmit} className="search-form">
           <input
